Cleanup and condition test execution
