+++ cs/src/core/FASTER.core.csproj
@@ -37,18 +37,4 @@
   <ItemGroup>
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.5.2" />
   </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Index/FASTER/Checkpoint.cs" />
-    <EmbeddedResource Include="Index/Common/Contexts.cs" />
-    <EmbeddedResource Include="Index/FASTER/FASTER.cs" />
-    <EmbeddedResource Include="Index/FASTER/FASTERBase.cs" />
-    <EmbeddedResource Include="Index/FASTER/FASTERImpl.cs" />
-    <EmbeddedResource Include="Index/FASTER/FASTERThread.cs" />
-    <EmbeddedResource Include="Index/FASTER/IndexCheckpoint.cs" />
-    <EmbeddedResource Include="Index/FASTER/IndexRecovery.cs" />
-    <EmbeddedResource Include="Index/FASTER/Recovery.cs" />
-    <EmbeddedResource Include="Utilities/AsyncResultTypes.cs" />
-    <EmbeddedResource Include="Utilities/StateTransitions.cs" />
-  </ItemGroup>
-
 </Project>
