+++ cs/src/devices/AzureStorageDevice/AzureStorageDevice.cs
@@ -11,108 +11,14 @@ using FASTER.core;
 using Microsoft.Azure.Storage;
 using Microsoft.Azure.Storage.Blob;
 
-namespace FASTER.cloud
+namespace FASTER.devices
 {
     /// <summary>
     /// A IDevice Implementation that is backed by<see href="https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-pageblob-overview">Azure Page Blob</see>.
-    /// This device is expected to be an order of magnitude slower than local SSD or HDD, but provide scalability and shared access in the cloud.
+    /// This device is slower than a local SSD or HDD, but provides scalability and shared access in the cloud.
     /// </summary>
-    public class AzurePageBlobDevice : StorageDeviceBase
+    public class AzureStorageDevice : StorageDeviceBase
     {
-        // This class bundles a page blob object with a queue and a counter to ensure 
-        // 1) BeginCreate is not called more than once
-        // 2) No writes are issued before EndCreate
-        // The creator of a BlobEntry is responsible for populating the object with an underlying Page Blob. Any subsequent callers
-        // either directly write to the created page blob, or queues the write so the creator can clear it after creation is complete.
-        // In-progress creation is denoted by a null value on the underlying page blob
-        private class BlobEntry
-        {
-            private CloudPageBlob pageBlob;
-            private ConcurrentQueue<Action<CloudPageBlob>> pendingWrites;
-            private int waitingCount;
-
-            /// <summary>
-            /// Creates a new BlobEntry, does not initialize a page blob. Use <see cref="CreateAsync(long, CloudPageBlob)"/>
-            /// for actual creation.
-            /// </summary>
-            public BlobEntry()
-            {
-                pageBlob = null;
-                pendingWrites = new ConcurrentQueue<Action<CloudPageBlob>>();
-                waitingCount = 0;
-            }
-
-            /// <summary>
-            /// Getter for the underlying <see cref="CloudPageBlob"/>
-            /// </summary>
-            /// <returns>the underlying <see cref="CloudPageBlob"/>, or null if there is none</returns>
-            public CloudPageBlob GetPageBlob()
-            {
-                return pageBlob;
-            }
-
-            /// <summary>
-            /// Asynchronously invoke create on the given pageBlob.
-            /// </summary>
-            /// <param name="size">maximum size of the blob</param>
-            /// <param name="pageBlob">The page blob to create</param>
-            public void CreateAsync(long size, CloudPageBlob pageBlob)
-            {
-                Debug.Assert(waitingCount == 0, "Create should be called on blobs that don't already exist and exactly once");
-                // Asynchronously create the blob
-                pageBlob.BeginCreate(size, ar =>
-                {
-                    try
-                    {
-                        pageBlob.EndCreate(ar);
-                    }
-                    catch (Exception e)
-                    {
-                        // TODO(Tianyu): Can't really do better without knowing error behavior
-                        Trace.TraceError(e.Message);
-                    }
-                    // At this point the blob is fully created. After this line all consequent writers will write immediately. We just
-                    // need to clear the queue of pending writers.
-                    this.pageBlob = pageBlob;
-                    // Take a snapshot of the current waiting count. Exactly this many actions will be cleared.
-                    // Swapping in -1 will inform any stragglers that we are not taking their actions and prompt them to retry (and call write directly)
-                    int waitingCountSnapshot = Interlocked.Exchange(ref waitingCount, -1);
-                    Action<CloudPageBlob> action;
-                    // Clear actions
-                    for (int i = 0; i < waitingCountSnapshot; i++)
-                    {
-                        // inserts into the queue may lag behind the creation thread. We have to wait until that happens.
-                        // This is so rare, that we are probably okay with a busy wait.
-                        while (!pendingWrites.TryDequeue(out action)) { }
-                        action(pageBlob);
-                    }
-                    // Mark for deallocation for the GC
-                    pendingWrites = null;
-                }, null);
-            }
-
-            /// <summary>
-            /// Attempts to enqueue an action to be invoked by the creator after creation is done. Should only be invoked when
-            /// creation is in-flight. This call is allowed to fail (and return false) if concurrently the creation is complete.
-            /// The caller should call the write action directly instead of queueing in this case.
-            /// </summary>
-            /// <param name="writeAction">The write action to perform</param>
-            /// <returns>Whether the action was successfully enqueued</returns>
-            public bool TryQueueAction(Action<CloudPageBlob> writeAction)
-            {
-                int currentCount;
-                do
-                {
-                    currentCount = waitingCount;
-                    // If current count became -1, creation is complete. New queue entries will not be processed and we must call the action ourselves.
-                    if (currentCount == -1) return false;
-                } while (Interlocked.CompareExchange(ref waitingCount, currentCount + 1, currentCount) != currentCount);
-                // Enqueue last. The creation thread is obliged to wait until it has processed waitingCount many actions.
-                // It is extremely unlikely that we will get scheduled out here anyways.
-                pendingWrites.Enqueue(writeAction);
-                return true;
-            }
-        }
         private CloudBlobContainer container;
         private readonly ConcurrentDictionary<int, BlobEntry> blobs;
         private readonly string blobName;
@@ -124,7 +30,7 @@ namespace FASTER.cloud
         private const uint PAGE_BLOB_SECTOR_SIZE = 512;
 
         /// <summary>
-        /// Constructs a new AzurePageBlobDevice instance
+        /// Constructs a new AzureStorageDevice instance, backed by Azure Page Blobs
         /// </summary>
         /// <param name="connectionString"> The connection string to use when estblishing connection to Azure Blobs</param>
         /// <param name="containerName">Name of the Azure Blob container to use. If there does not exist a container with the supplied name, one is created</param>
@@ -133,7 +39,7 @@ namespace FASTER.cloud
         /// True if the program should delete all blobs created on call to <see cref="Close">Close</see>. False otherwise. 
         /// The container is not deleted even if it was created in this constructor
         /// </param>
-        public AzurePageBlobDevice(string connectionString, string containerName, string blobName, bool deleteOnClose = false)
+        public AzureStorageDevice(string connectionString, string containerName, string blobName, bool deleteOnClose = false)
             : base(connectionString + "/" + containerName + "/" + blobName, PAGE_BLOB_SECTOR_SIZE)
         {
             CloudStorageAccount storageAccount = CloudStorageAccount.Parse(connectionString);
