+++ cs/src/devices/AzureStorageDevice/FASTER.devices.AzureStorageDevice.nuspec
@@ -1,0 +1,34 @@
+<?xml version="1.0"?>
+<package >
+  <metadata>
+    <id>FASTER.devices.AzureStorageDevice</id>
+    <version>$version$</version>
+    <title>FASTER.devices.AzureStorageDevice</title>
+    <authors>Microsoft</authors>
+    <owners>Microsoft</owners>
+    <projectUrl>https://github.com/Microsoft/FASTER</projectUrl>
+    <licenseUrl>https://raw.githubusercontent.com/Microsoft/FASTER/master/LICENSE</licenseUrl>
+    <requireLicenseAcceptance>true</requireLicenseAcceptance>
+    <description>IDevice FASTER is a fast concurrent key-value store that also supports indexing of larger-than-memory data. This is a FASTER IDevice implementation for Azure Storage.</description>
+    <releaseNotes>See the project website at https://github.com/Microsoft/FASTER for more details</releaseNotes>
+    <copyright>© Microsoft Corporation. All rights reserved.</copyright>
+    <language>en-US</language>
+    <tags>key-value store dictionary hashtable concurrent log persistent azure storage FASTER</tags>
+    <dependencies>
+      <group targetFramework="net46">
+        <dependency id="Microsoft.Azure.Storage.Blob" version="10.0.3" />
+      </group>
+      <group targetFramework="netstandard2.0">
+        <dependency id="Microsoft.Azure.Storage.Blob" version="10.0.3" />
+      </group>
+    </dependencies>
+  </metadata>
+  <files>
+    <file src="bin\AnyCPU\Release\net46\FASTER.devices.AzureStorageDevice.dll" target="lib\net46" />
+    <file src="bin\AnyCPU\Release\net46\FASTER.devices.AzureStorageDevice.pdb" target="lib\net46" />
+    <file src="bin\AnyCPU\Release\net46\FASTER.devices.AzureStorageDevice.xml" target="lib\net46" />
+    <file src="bin\AnyCPU\Release\netstandard2.0\FASTER.devices.AzureStorageDevice.dll" target="lib\netstandard2.0" />
+    <file src="bin\AnyCPU\Release\netstandard2.0\FASTER.devices.AzureStorageDevice.pdb" target="lib\netstandard2.0" />
+    <file src="bin\AnyCPU\Release\netstandard2.0\FASTER.devices.AzureStorageDevice.xml" target="lib\netstandard2.0" />
+  </files>
+</package>
