+++ cs/test/FASTER.test.csproj
@@ -43,7 +43,7 @@
     <PackageReference Include="NUnit3TestAdapter" Version="3.10.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\src\cloud\FASTER.cloud.csproj" />
     <ProjectReference Include="..\src\core\FASTER.core.csproj" />
+    <ProjectReference Include="..\src\devices\AzureStorageDevice\FASTER.devices.AzureStorageDevice.csproj" />
   </ItemGroup>
 </Project>
